Refactor MoveSearchForm into a new component
